+++ Sources/Guppy.swift
@@ -16,6 +16,9 @@
 //  limitations under the License.
 //
 
+import Foundation
+import UIKit
+
 public class Guppy {
 
     // MARK: - Public
@@ -27,14 +30,6 @@ public class Guppy {
     /// Registers custom URLProtocol. If you are you using a custom URLSession add **GuppyURLProtocol.self** to your URLSessionConfiguration's protocolClasses
     public static func registerURLProtocol() {
         URLProtocol.registerClass(GuppyURLProtocol.self)
-    }
-
-    /// Using configuration.protocolClasses.append(getSniffURLProtocol().self) might not work if you have a URLProtocol already in the protocolClass array. It is recommend to create a new AnyClass array and set it to protocolClasses.
-    ///
-    /// - Returns: Custom url protocol as AnyClass and can be added to a URLSessionConfiguration's protocol classes.
-    @available(*, deprecated, message: "Please use `GuppyURLProtocol.self` directly")
-    public static func getSniffURLProtocol() -> AnyClass {
-        return GuppyURLProtocol.self
     }
 
     public func getAllLogs() -> [LogItem] {